Make only intended charts draggable
On a page with other charts, this plugin seemed to reduce the chart padding with the `chart.redraw()` line, and made columns have a dotted border if the values were near 0 (which should happen on draggable charts).  I think if you want to have a draggable chart, you shouldn't have to have these events and styles affect other charts.

+++ draggable-points.js
@@ -32,8 +32,15 @@
             dragX,
             dragY,
             dragPlotX,
-            dragPlotY;
-
+            dragPlotY,
+			draggable = chart.options.chart.draggable;
+        
+        // set chart.draggable = true to enable dragging
+        if (!draggable) {
+            return false;
+        }
+        
+        // this seems to cause reduced padding on other charts
         chart.redraw(); // kill animation (why was this again?)
 
         addEvent(container, 'mousedown', function (e) {
@@ -134,7 +141,11 @@
         baseDrawTracker = colProto.drawTracker;
 
     colProto.drawTracker = function () {
-        var series = this;
+        var series = this,
+			userOptions = series.userOptions;
+		if (!userOptions.draggableY && !userOptions.draggableX) {
+			return false; // don't change non-draggable charts
+		}
         baseDrawTracker.apply(series);
         each(series.points, function (point) {
             point.graphic.attr(point.shapeArgs.height < 3 ? {
